Refactor populate.ts and schema files for improved code organization and type consistency
